N°6208 - Router: Move to Service folder
